Finished adding data, but with the following:
Ward2 and WardDB2 are the old way Ward was structured, and Ward/WardDB
are the way that matches demo data. Old way is also commented out in the
table generator.

Removed BillingRecordCost table bc it's useless.

For some reason checkin data isn't getting added to the db. All other
demo data is, but this isn't and I am wondering if it's bc
end_date=null.