Changed it to use the school servers and added the mariadb jdbc
The only caveat is that if you are using it on a computer that is not
connected to NCSU's network then you will need to use a vpn.

It uses my login for the database but I don't mind sharing the password
for it since I changed it to be a generic password I don't use anywhere
else.